% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bash.R
\name{bash}
\alias{bash}
\title{Take a Bash}
\usage{
bash(expr, e = NULL, w = NULL)
}
\arguments{
\item{expr}{Expression to run, passed to \code{\link[=tryCatch]{tryCatch()}}.}

\item{e, w}{An object of class \code{Error} or \code{Warning} as
returned by \code{\link[=e]{e()}} or \code{\link[=w]{w()}}.}
}
\description{
Equivalent to \code{\link[=tryCatch]{tryCatch()}}.
}
\examples{
safe_log <- function(x){
 result <- bash(log(x))
 
 if(is.e(result))
   stop(result$stop())

 return(result)
} 

if(interactive())
 safe_log("a")

}
