% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{checks}
\alias{checks}
\alias{is.e}
\alias{is.w}
\title{Check}
\usage{
is.e(e)

is.w(w)
}
\arguments{
\item{e, w}{Object to check.}
}
\description{
Check whether an object is an error or a warning.
}
\examples{
err <- e("Whoops!")

is.e(err)
is.w(err)

}
