% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enforce.R
\name{chk}
\alias{chk}
\alias{chk.default}
\alias{chk.err}
\title{Check}
\usage{
chk(obj)

\method{chk}{default}(obj)

\method{chk}{err}(obj)
}
\arguments{
\item{obj}{Object to check.}
}
\description{
Checks individual objects.
}
\details{
Runs \code{\link[=warning]{warning()}} or \code{\link[=stop]{stop()}} where necessary.
}
