% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enforce.R
\name{enforces}
\alias{enforces}
\alias{enforce}
\alias{defer_enforce}
\title{Enforce Errors and Warnings}
\usage{
enforce(...)

defer_enforce(...)
}
\arguments{
\item{...}{Objects to check, if any of them is an \code{Error}
then \code{\link[=stop]{stop()}} is called, if any are \code{Warning}s then \code{\link[=warning]{warning()}}
is called.}
}
\description{
Enforce Errors and Warnings
}
\details{
Objects passed are evalutated in order.
}
