% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/err.R
\name{err}
\alias{err}
\alias{error}
\alias{error.character}
\alias{error.error}
\alias{warn}
\alias{warn.character}
\alias{warn.warning}
\title{Error & Warnings}
\usage{
error(error)

\method{error}{character}(error)

\method{error}{error}(error)

warn(warning)

\method{warn}{character}(warning)

\method{warn}{warning}(warning)
}
\arguments{
\item{error, warning}{Message string, object of class \code{error},
or \code{warning}.}
}
\description{
Create new errors and warnings.
}
