% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{Error}
\alias{Error}
\title{Error}
\description{
Error

Error
}
\section{Super class}{
\code{\link[erratum:Issue]{erratum::Issue}} -> \code{Error}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Error-new}{\code{Error$new()}}
\item \href{#method-Error-stop}{\code{Error$stop()}}
\item \href{#method-Error-fatal}{\code{Error$fatal()}}
\item \href{#method-Error-clone}{\code{Error$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="erratum" data-topic="Issue" data-id="addRule"><a href='../../erratum/html/Issue.html#method-Issue-addRule'><code>erratum::Issue$addRule()</code></a></span></li>
<li><span class="pkg-link" data-pkg="erratum" data-topic="Issue" data-id="check"><a href='../../erratum/html/Issue.html#method-Issue-check'><code>erratum::Issue$check()</code></a></span></li>
<li><span class="pkg-link" data-pkg="erratum" data-topic="Issue" data-id="print"><a href='../../erratum/html/Issue.html#method-Issue-print'><code>erratum::Issue$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="erratum" data-topic="Issue" data-id="raise"><a href='../../erratum/html/Issue.html#method-Issue-raise'><code>erratum::Issue$raise()</code></a></span></li>
<li><span class="pkg-link" data-pkg="erratum" data-topic="Issue" data-id="return"><a href='../../erratum/html/Issue.html#method-Issue-return'><code>erratum::Issue$return()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Error-new"></a>}}
\if{latex}{\out{\hypertarget{method-Error-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Error$new(obj, raiser = getOption("ERR_RAISER_ERROR", stopper))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj}}{A character string or an object of
class \code{error}, or \code{warning}.}

\item{\code{raiser}}{Template to raise the issue.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Initialise
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Error-stop"></a>}}
\if{latex}{\out{\hypertarget{method-Error-stop}{}}}
\subsection{Method \code{stop()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Error$stop()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Stop

Analogous to \code{\link[=stop]{stop()}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Error-fatal"></a>}}
\if{latex}{\out{\hypertarget{method-Error-fatal}{}}}
\subsection{Method \code{fatal()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Error$fatal()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Fatal

Analogous to \code{\link[=stop]{stop()}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Error-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Error-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Error$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
