% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{get_call}
\alias{get_call}
\title{Extract Call}
\usage{
get_call(obj)
}
\arguments{
\item{obj}{Message string, object of class \code{error},
or \code{warning}.}
}
\description{
Extract call from error and warnings.
}
