% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{get_msg}
\alias{get_msg}
\title{Extract Message}
\usage{
get_msg(obj)
}
\arguments{
\item{obj}{Message string, object of class \code{error},
or \code{warning}.}
}
\description{
Extract message from error and warnings.
}
