% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issue.R
\name{Issue}
\alias{Issue}
\title{Core Class}
\description{
Core class to create and handle issues.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Issue$new()}}
\item \href{#method-message}{\code{Issue$message()}}
\item \href{#method-print}{\code{Issue$print()}}
\item \href{#method-return}{\code{Issue$return()}}
\item \href{#method-clone}{\code{Issue$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Issue$new(obj, type = c("error", "warning"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj}}{A character string or an object of
class \code{error}, or \code{warning}.}

\item{\code{type}}{Type of message.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Initialise
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-message"></a>}}
\if{latex}{\out{\hypertarget{method-message}{}}}
\subsection{Method \code{message()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Issue$message()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Message

Retrieve the message
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Issue$print()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Print

Print message of error or warning.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-return"></a>}}
\if{latex}{\out{\hypertarget{method-return}{}}}
\subsection{Method \code{return()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Issue$return()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Return

Returns self from parent function.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Issue$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
