% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/err.R
\name{new}
\alias{new}
\title{Error & Warnings}
\usage{
new(error, ..., type = c("error", "warning"))
}
\arguments{
\item{error}{Message string, object of class \code{error},
or \code{warning}.}

\item{...}{Ignored.}

\item{type}{Type of the message \code{error} or \code{warning}.}
}
\description{
Create new errors and warnings.
}
