% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{raise}
\alias{raise}
\alias{raise.e}
\alias{raise.w}
\title{Raisers}
\usage{
raise.e(fn = NULL)

raise.w(fn = NULL)
}
\arguments{
\item{fn}{Function to run when the \code{raise} method
is called. By default the error uses \code{stop()} and warning
uses \code{warning()}. The function must accept a single argument:
the error message (character vector).}
}
\description{
Set \code{raise} method globally, every subsequent \code{raise}
method will make use of this function.
}
