% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve.R
\name{resolves}
\alias{resolves}
\alias{resolve}
\alias{defer_resolve}
\alias{enforce}
\alias{defer_enforce}
\title{Resolve Errors and Warnings}
\usage{
resolve(...)

defer_resolve(...)

enforce(...)

defer_enforce(...)
}
\arguments{
\item{...}{Objects to check, if any of them is an \code{Error}
then \code{\link[=stop]{stop()}} is called, if any are \code{Warning}s then \code{\link[=warning]{warning()}}
is called.}
}
\value{
Invisiby returns \code{NULL}
}
\description{
Resolve Errors and Warnings
}
\details{
Objects passed are evalutated in order.
}
