% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skip.R
\name{skip}
\alias{skip}
\title{Skip}
\usage{
skip(..., w = FALSE)
}
\arguments{
\item{...}{Objects to check, if any of them is an \code{Error}
then it calls \code{\link[=return]{return()}} in the parent function, this can
optionally be applied if any object is a \code{Warning} with the
\code{w} argument.}

\item{w}{Whether to also skip is there are \code{Warning}.}
}
\description{
Skip the rest of the function; calls \code{\link[=return]{return()}} in the
parent function if any object is an error or (optionally)
a warning.
}
