% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{template}
\alias{template}
\alias{tmpl.e}
\alias{tmpl.w}
\title{Templates}
\usage{
tmpl.e(pat)

tmpl.w(pat)
}
\arguments{
\item{pat}{Pattern to use, must include \verb{\%s},
forwarded to \code{\link[=sprintf]{sprintf()}}.}
}
\description{
Define error and warning templates.
}
\examples{
msg <- "Something's wrong" 

# default
e(msg)

# template
tmpl.e("Whoops: \%s - sorry!")
e(msg)


}
