% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{tmpl}
\alias{tmpl}
\title{Fetch Default Templates}
\usage{
tmpl(type = c("error", "warning"))
}
\arguments{
\item{type}{What type of template to fetch.}
}
\description{
Fetch Default Templates
}
\keyword{internal}
