% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{checks}
\alias{checks}
\alias{is.e}
\alias{is.e.default}
\alias{is.e.err}
\alias{is.w}
\alias{is.w.default}
\alias{is.w.err}
\title{Check}
\usage{
is.e(obj)

\method{is.e}{default}(obj)

\method{is.e}{err}(obj)

is.w(obj)

\method{is.w}{default}(obj)

\method{is.w}{err}(obj)
}
\arguments{
\item{obj}{Object to check.}
}
\description{
Check whether an object is an error or a warning.
}
\examples{
err <- e("Whoops!")

is.e(err)
is.w(err)

}
