% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{Error}
\alias{Error}
\title{Error}
\description{
Error

Error
}
\section{Super class}{
\code{\link[err:Issue]{err::Issue}} -> \code{Error}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-stop}{\code{Error$stop()}}
\item \href{#method-clone}{\code{Error$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="err" data-topic="Issue" data-id="initialize">}\href{../../err/html/Issue.html#method-initialize}{\code{err::Issue$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="err" data-topic="Issue" data-id="message">}\href{../../err/html/Issue.html#method-message}{\code{err::Issue$message()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="err" data-topic="Issue" data-id="print">}\href{../../err/html/Issue.html#method-print}{\code{err::Issue$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stop"></a>}}
\if{latex}{\out{\hypertarget{method-stop}{}}}
\subsection{Method \code{stop()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Error$stop()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Stop

Analogous to \code{\link[=stop]{stop()}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Error$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
