% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{Error}
\alias{Error}
\title{Error}
\description{
Error

Error
}
\section{Super class}{
\code{\link[erratum:Issue]{erratum::Issue}} -> \code{Error}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-stop}{\code{Error$stop()}}
\item \href{#method-fatal}{\code{Error$fatal()}}
\item \href{#method-clone}{\code{Error$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="erratum" data-topic="Issue" data-id="addRule">}\href{../../erratum/html/Issue.html#method-addRule}{\code{erratum::Issue$addRule()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="erratum" data-topic="Issue" data-id="check">}\href{../../erratum/html/Issue.html#method-check}{\code{erratum::Issue$check()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="erratum" data-topic="Issue" data-id="initialize">}\href{../../erratum/html/Issue.html#method-initialize}{\code{erratum::Issue$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="erratum" data-topic="Issue" data-id="print">}\href{../../erratum/html/Issue.html#method-print}{\code{erratum::Issue$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="erratum" data-topic="Issue" data-id="raise">}\href{../../erratum/html/Issue.html#method-raise}{\code{erratum::Issue$raise()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="erratum" data-topic="Issue" data-id="return">}\href{../../erratum/html/Issue.html#method-return}{\code{erratum::Issue$return()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stop"></a>}}
\if{latex}{\out{\hypertarget{method-stop}{}}}
\subsection{Method \code{stop()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Error$stop()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Stop

Analogous to \code{\link[=stop]{stop()}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fatal"></a>}}
\if{latex}{\out{\hypertarget{method-fatal}{}}}
\subsection{Method \code{fatal()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Error$fatal()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Fatal

Analogous to \code{\link[=stop]{stop()}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Error$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
