% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{template}
\alias{template}
\alias{template.e}
\alias{template.w}
\title{Templates}
\usage{
template.e(pat = "\%s")

template.w(pat = "\%s")
}
\arguments{
\item{pat}{Pattern to use, must include \verb{\%s},
forwarded to \code{\link[=sprintf]{sprintf()}}.}
}
\description{
Define error and warning templates.
}
\examples{
msg <- "Something's wrong" 

# default
e(msg)

# template
template.e("Whoops: \%s - sorry!")
e(msg)

# reset
template.e()

}
