% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warning.R
\name{Warning}
\alias{Warning}
\title{Error}
\description{
Error

Error
}
\section{Super class}{
\code{\link[err:Issue]{err::Issue}} -> \code{Warning}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-warning}{\code{Warning$warning()}}
\item \href{#method-clone}{\code{Warning$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="err" data-topic="Issue" data-id="initialize">}\href{../../err/html/Issue.html#method-initialize}{\code{err::Issue$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="err" data-topic="Issue" data-id="print">}\href{../../err/html/Issue.html#method-print}{\code{err::Issue$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-warning"></a>}}
\if{latex}{\out{\hypertarget{method-warning}{}}}
\subsection{Method \code{warning()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Warning$warning()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Display the warning message
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Warning$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
