% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warning.R
\name{Warning}
\alias{Warning}
\title{Error}
\description{
Error

Error
}
\section{Super class}{
\code{\link[erratum:Issue]{erratum::Issue}} -> \code{Warning}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Warning-new}{\code{Warning$new()}}
\item \href{#method-Warning-warn}{\code{Warning$warn()}}
\item \href{#method-Warning-clone}{\code{Warning$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="erratum" data-topic="Issue" data-id="addRule"><a href='../../erratum/html/Issue.html#method-Issue-addRule'><code>erratum::Issue$addRule()</code></a></span></li>
<li><span class="pkg-link" data-pkg="erratum" data-topic="Issue" data-id="check"><a href='../../erratum/html/Issue.html#method-Issue-check'><code>erratum::Issue$check()</code></a></span></li>
<li><span class="pkg-link" data-pkg="erratum" data-topic="Issue" data-id="print"><a href='../../erratum/html/Issue.html#method-Issue-print'><code>erratum::Issue$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="erratum" data-topic="Issue" data-id="raise"><a href='../../erratum/html/Issue.html#method-Issue-raise'><code>erratum::Issue$raise()</code></a></span></li>
<li><span class="pkg-link" data-pkg="erratum" data-topic="Issue" data-id="return"><a href='../../erratum/html/Issue.html#method-Issue-return'><code>erratum::Issue$return()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Warning-new"></a>}}
\if{latex}{\out{\hypertarget{method-Warning-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Warning$new(obj, raiser = getOption("ERR_RAISER_WARNING", warner))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj}}{A character string or an object of
class \code{error}, or \code{warning}.}

\item{\code{raiser}}{Template to raise the issue.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Initialise
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Warning-warn"></a>}}
\if{latex}{\out{\hypertarget{method-Warning-warn}{}}}
\subsection{Method \code{warn()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Warning$warn()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Warn

Analogous to \code{\link[=warning]{warning()}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Warning-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Warning-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Warning$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
