% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warning.R
\name{Warning}
\alias{Warning}
\title{Error}
\description{
Error

Error
}
\section{Super class}{
\code{\link[erratum:Issue]{erratum::Issue}} -> \code{Warning}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-warn}{\code{Warning$warn()}}
\item \href{#method-clone}{\code{Warning$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="erratum" data-topic="Issue" data-id="addRule">}\href{../../erratum/html/Issue.html#method-addRule}{\code{erratum::Issue$addRule()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="erratum" data-topic="Issue" data-id="check">}\href{../../erratum/html/Issue.html#method-check}{\code{erratum::Issue$check()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="erratum" data-topic="Issue" data-id="initialize">}\href{../../erratum/html/Issue.html#method-initialize}{\code{erratum::Issue$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="erratum" data-topic="Issue" data-id="print">}\href{../../erratum/html/Issue.html#method-print}{\code{erratum::Issue$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="erratum" data-topic="Issue" data-id="raise">}\href{../../erratum/html/Issue.html#method-raise}{\code{erratum::Issue$raise()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="erratum" data-topic="Issue" data-id="return">}\href{../../erratum/html/Issue.html#method-return}{\code{erratum::Issue$return()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-warn"></a>}}
\if{latex}{\out{\hypertarget{method-warn}{}}}
\subsection{Method \code{warn()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Warning$warn()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Warn

Analogous to \code{\link[=warning]{warning()}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Warning$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
